+++ src/rendering/RenderPass.cpp
@@ -40,7 +40,7 @@ RenderPass::RenderPass(const Array<VkFormat>& formats, bool present) : _formats(
 
   VkRenderPassCreateInfo render_pass_info = {};
   render_pass_info.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-  render_pass_info.attachmentCount = _attachment_descriptions.size();
+  render_pass_info.attachmentCount = uint32_t(_attachment_descriptions.size());
   render_pass_info.pAttachments = _attachment_descriptions.begin();
   render_pass_info.subpassCount = 1;
   render_pass_info.pSubpasses = &subpass_desc;
